src_analytics_api/pyproject.toml
[project]
name         = "analytics_api"
description  = "Internet of Things Web API"
authors      = [{name = "Joshoua Bigler"}]
version      = "0.0.1"
dependencies = [
  "fastapi>=0.115.6",
  "mlflow>=2.21.3",
  "python-dotenv>=1.0.1",
  "uvicorn>=0.31.1",
  "strawberry-graphql>=0.261.1",
  "torch>=2.6.0",
  "iot-libs>=0.0.9"
]

[tool.setuptools]
package-dir  = { "" = "." }
packages = { find = { include = ["analytics_api*"] } } 
[tool.pytest.ini_options]
testpaths = ["test"]